업그레이드 매니저 마저 작성 / 스프라이트 정리 / HP 시스템 float으로 변경

+++ Dodge missile/Assets/02.Scripts/Health.cs
@@ -4,16 +4,16 @@ using UnityEngine;
 
 public abstract class Health : MonoBehaviour, IHitable
 {
-    [SerializeField] int _maxHp = 30;
+    [SerializeField] float _maxHp = 30;
 
-    protected int _hp = 0;
+    protected float _hp = 0;
 
     protected virtual void Awake()
     {
         _hp = _maxHp;
     }
 
-    public virtual void OnHit(int damage)
+    public virtual void OnHit(float damage)
     {
         if (!gameObject.activeSelf)
             return;
